Expose renderer base classes
