+++ plotly/io/__init__.py
@@ -6,3 +6,5 @@ from ._json import to_json, from_json, read_json, write_json
 from ._templates import templates, to_templated
 
 from ._renderers import renderers, show
+
+from . import base_renderers
