+++ plotly/io/base_renderers.py
@@ -1,0 +1,5 @@
+from ._base_renderers import (
+    MimetypeRenderer, PlotlyRenderer, JsonRenderer, ImageRenderer,
+    PngRenderer, SvgRenderer, PdfRenderer, JpegRenderer, HtmlRenderer,
+    ColabRenderer, KaggleRenderer, NotebookRenderer, SideEffectRenderer,
+    BrowserRenderer)
